+++ examples/react-native/expo-app/App.tsx
@@ -52,8 +52,11 @@ export default function App() {
     (async () => {
         // AMPLITUDE_API_KEY is inlined at bundle time (see babel.config.js).
         await init(process.env.AMPLITUDE_API_KEY || 'YOUR_API_KEY', 'react-native-user-id', {
-          logLevel: Types.LogLevel.Error,
-          // autocapture: { } // <-- todo
+          logLevel: Types.LogLevel.Debug,
+          autocapture: {
+            sessions: true,
+            appLifecycles: true,
+          } // <-- todo
         }).promise;
 
         // TODO: This should be replaced with the Autocapture Plugin