+++ packages/analytics-core/src/index.ts
@@ -152,6 +152,8 @@ export { NodeConfig, NodeOptions } from './types/config/node-config';
 // React Native
 export {
   ReactNativeConfig,
+  ReactNativeConfigAutocaptureBeta,
+  ReactNativeAutocaptureOptions,
   ReactNativeTrackingOptions,
   ReactNativeOptions,
   ReactNativeAttributionOptions,
@@ -180,3 +182,4 @@ export { ExcludeInternalReferrersOptions, EXCLUDE_INTERNAL_REFERRERS_CONDITIONS 
 export { VideoObserver, State as VideoState, type VideoObserverParams } from './observers/video';
 export { EmbeddedVideoPlayer, type Vendor as VideoVendor } from './video-analytics/types';
 export { isChromeExtension, isReactNative } from './utils/environment';
+export { translateRemoteConfigToLocal } from './config/joined-config';