+++ packages/analytics-react-native/src/index.ts
@@ -25,7 +25,7 @@ export const {
 
 export { Revenue, Identify } from '@amplitude/analytics-core';
 
-export { ampCapture, subscribe } from './amp-capture';
+export { ampCapture } from './amp-capture';
 
 // Export types to maintain backward compatibility with `analytics-types`.
 // In the next major version, only export customer-facing types to reduce the public API surface.