+++ packages/analytics-node/CHANGELOG.md
@@ -3,6 +3,44 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.66](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.65...@amplitude/analytics-node@1.5.66) (2026-07-17)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
+## [1.5.65](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.64...@amplitude/analytics-node@1.5.65) (2026-07-14)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
+## [1.5.64](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.63...@amplitude/analytics-node@1.5.64) (2026-07-13)
+
+
+### Bug Fixes
+
+* **analytics-core:** queued functions not being flushed when init not awaited ([#1873](https://github.com/amplitude/Amplitude-TypeScript/issues/1873)) ([e0931a1](https://github.com/amplitude/Amplitude-TypeScript/commit/e0931a132e40c9f72a9bab87fd7f578457bc6c8f))
+
+
+
+
+
+## [1.5.63](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.62...@amplitude/analytics-node@1.5.63) (2026-07-09)
+
+
+### Bug Fixes
+
+* **analytics-node:** export NodeClient and NodeOptions from Types namespace ([#1759](https://github.com/amplitude/Amplitude-TypeScript/issues/1759)) ([820ef6f](https://github.com/amplitude/Amplitude-TypeScript/commit/820ef6fe1ce73a08e244ffaef86c3183c69c9be6))
+
+
+
+
+
 ## [1.5.62](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.61...@amplitude/analytics-node@1.5.62) (2026-06-30)
 
 **Note:** Version bump only for package @amplitude/analytics-node