+++ packages/session-replay-browser/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.47.4](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.47.3...@amplitude/session-replay-browser@1.47.4) (2026-07-17)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
+## [1.47.3](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.47.2...@amplitude/session-replay-browser@1.47.3) (2026-07-14)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
+## [1.47.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.47.1...@amplitude/session-replay-browser@1.47.2) (2026-07-13)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
+## [1.47.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.47.0...@amplitude/session-replay-browser@1.47.1) (2026-07-09)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
 # [1.47.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.46.2...@amplitude/session-replay-browser@1.47.0) (2026-06-30)